test: Pio vs FwのE/P Test
